api-server/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.0.0

# HTTP client (async)
httpx>=0.27.0

# PDF processing
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
Pillow>=10.0.0

# Word / PowerPoint (optional)
python-docx>=1.1.0
python-pptx>=1.0.0

# File upload
python-multipart>=0.0.9
